fix: Split out tests
